% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendseason.R
\name{friedman.test}
\alias{friedman.test}
\title{Friedman (JASA) Test for Presence of Seasonality}
\usage{
friedman.test(y,r)
}
\arguments{
\item{y}{is vector representing the observed time series with n observations.}

\item{r}{is a positive integer representing the no. of seasons.}
}
\value{
Results of the Freidman test.
}
\description{
Testing for Presence of Seasonality using Friedman (JASA) Test.
}
\examples{
friedman.test(AirPassengers,12)
}
\author{
Abhishek K. Umrawal
}
\seealso{
\code{rank}
}
