% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendseason.R
\name{ro.test}
\alias{ro.test}
\title{Relative Ordering Test for Presence of Trend}
\usage{
ro.test(y)
}
\arguments{
\item{y}{is vector representing the observed time series with n observations.}
}
\value{
Results of the relative ordering test.
}
\description{
Testing for Presence of Trend using Relative Ordering Test.
}
\examples{
ro.test(AirPassengers)
}
\author{
Abhishek K. Umrawal
}
\seealso{
\code{rank}
}

